Before refactoring the db
