+++ .gitignore
@@ -5,4 +5,4 @@
 /captures
 *.iml
 /.idea
-/gradle 
+/gradle
