+++ app/src/main/java/com/szymkowski/personaltrainercompanion/payments/domain/PaymentMapper.java
@@ -7,11 +7,6 @@ import org.mapstruct.ReportingPolicy;
 import org.mapstruct.factory.Mappers;
 
 
-/*
-This is probably the biggest overkill ever, to use mapstruct just for this one mapper.
-The heck, I wanted to learn how to do this.
- */
-
 @Mapper(unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface PaymentMapper {
