+++ app/src/main/java/com/szymkowski/personaltrainercompanion/trainings/providers/PaidNumberOfTrainingsProvider.java
@@ -1,4 +1,4 @@
-package com.szymkowski.personaltrainercompanion.trainings.domain;
+package com.szymkowski.personaltrainercompanion.trainings.providers;
 
 public interface PaidNumberOfTrainingsProvider {
     int getNumberOfTrainingsPaidFor();
